extend StructureContainer instead of new class
StructureList renamed to StructureContainer, it also behaves like a
atomistic job now by forwarding `self.structure` to `self.structures[0]`
