+++ pyiron/atomistics/job/structurecontainer.py
@@ -2,62 +2,160 @@
 # Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
 # Distributed under the terms of "New BSD License", see the LICENSE file.
 
-from __future__ import print_function
-from pyiron.atomistics.structure.atoms import Atoms
-from pyiron.atomistics.job.atomistic import AtomisticGenericJob
-
-__author__ = "Yury Lysogorskiy"
+__author__ = "Marvin Poul"
 __copyright__ = (
     "Copyright 2020, Max-Planck-Institut für Eisenforschung GmbH - "
     "Computational Materials Design (CM) Department"
 )
-__version__ = "1.0"
-__maintainer__ = "Jan Janssen"
-__email__ = "janssen@mpie.de"
+__version__ = "0.1"
+__maintainer__ = "Marvin Poul"
+__email__ = "poul@mpie.de"
 __status__ = "development"
-__date__ = "Sep 1, 2017"
+__date__ = "Aug 12, 2020"
 
+from pyiron.base.generic.inputlist import InputList
+from pyiron.atomistics.job.atomistic import AtomisticGenericJob
+from pyiron.atomistics.structure.atoms import Atoms
 
-class StructureContainer(AtomisticGenericJob):
+class StructureList(AtomisticGenericJob):
+    """
+    Container to save a list of structures in HDF5 together with tags.
+
+    Add new structures with :meth:`.StructureList.append`, they are
+    added to :attr:`.StructureList.structures`.  The HDF5 is written when
+    :meth:`.run` is called.
+    """
+
+    __version__ = "0.1.0"
+
     def __init__(self, project, job_name):
-        super(StructureContainer, self).__init__(project, job_name)
-        self.__version__ = "0.1"
-        self.__name__ = "StructureContainer"
+        super().__init__(project, job_name)
+        self.__structures = InputList(table_name = "structures")
         self.server.run_mode.interactive = True
 
     @property
+    def structures(self):
+        """
+        :class:`.InputList`: list of structures
+        with meta data; each item in this list has an 'atoms' key for the
+        atomic structures and then as many additional keys as passed to
+        :meth:`~pyiron.
+        """
+        return self.__structures
+
+    @staticmethod
+    def __to_structure(structure_or_job):
+        """
+        If :class:`~.AtomisticGenericJob` try to get most recent structure, if
+        :class:`~.Atoms` return as is, throw an error otherwise.  Wrap
+        structures in dict under the key "atoms".
+        """
+        if isinstance(structure_or_job, AtomisticGenericJob):
+            if structure.structure:
+                return {
+                        "atoms": structure_or_job.get_structure(iteration_step=-1),
+                        "jobid": structure_or_job.job_id
+                }
+            else:
+                raise ValueError(
+                        "The job does not contain any structure to import."
+                )
+        elif isinstance(structure_or_job, Atoms):
+            return {"atoms": structure_or_job}
+        else:
+            raise TypeError(
+                "You can only use a structure object or an "
+                "AtomisticGenericJob object."
+            )
+
+    @property
     def structure(self):
-        return self._structure
+        """
+        :class:`~.Atoms`: first (or only) structure set in the container
+        """
+        return self.structures[0].atoms
 
     @structure.setter
     def structure(self, structure):
-        if isinstance(structure, AtomisticGenericJob):
-            if structure.structure:
-                self._structure = structure.get_structure(iteration_step=-1)
-                self.parent_id = structure.job_id
-            else:
-                raise ValueError("The job does not contain any structure to import.")
-        elif isinstance(structure, Atoms):
-            self._structure = structure
+        item = self.__to_structure(structure_or_job)
+        if len(self.structures) >= 1:
+            self.structures[0] = item
         else:
-            raise TypeError(
-                "You can only append a structure object or an GenericJob object."
-            )
-        if self.status.initialized:
-            self.run()
+            self.structures.append(item)
 
-    def append(self, structure_to_append):
-        self.structure = structure_to_append
+    def append(self, structure_or_job, **kwargs):
+        """
+        Add new structure to structure list.
 
-    def run_if_interactive(self):
+        The added structure will available in
+        :attr:`~.StructureList.structures`.  If **kwargs contains a key "atoms"
+        is ignored, since the structure is saved under this key.  If the
+        structure is added via a job, retrieve the latest structure and its id
+        is also saved under the "jobid" key and **kwargs of the same name are
+        ignored.
+
+        Args:
+            structure_or_job (AtomisticGenericJob/Atoms):
+                if :class:`~.AtomisticGenericJob` add from
+                :meth:`~.AtomisticGenericJob.get_structure`,
+                otherwise add just the given :class:`~.Atoms`
+            **kwargs: meta data tags to add to the structure in the list
+
+        Returns:
+            dict: item added to :attr:`~.structures`
+        """
+
+        kwargs.update(self.__to_structure(structure_or_job))
+        self.__structures.append(kwargs)
+
+        return self.structures[0]
+
+    def run_static(self):
         self.status.finished = True
 
-    def to_hdf(self, hdf=None, group_name=None):
-        super(StructureContainer, self).to_hdf(hdf=hdf, group_name=group_name)
-        with self.project_hdf5.open("input") as hdf5_input:
-            self.structure.to_hdf(hdf5_input)
+    def run_if_interactive(self):
+        self.to_hdf()
+        self.status.finished = True
 
-    def from_hdf(self, hdf=None, group_name=None):
-        super(StructureContainer, self).from_hdf(hdf=hdf, group_name=group_name)
-        with self.project_hdf5.open("input") as hdf5_input:
-            self.structure = Atoms().from_hdf(hdf5_input)
+    def write_input(self):
+        pass
+
+    def collect_output(self):
+        pass
+
+    def to_hdf(self, hdf = None, group_name = None):
+        # skip any of the AtomisticGenericJob specific serialization, since we
+        # handle the structures on our own and that method might just confuse
+        # self.structure and self.structures
+        super(GenericJob, self).to_hdf(hdf = hdf, group_name = group_name)
+
+        hdf = self.project_hdf5.create_group("structures")
+
+        for i, s in enumerate(self.__structures.values()):
+            sub = hdf.create_group("index_{}".format(i))
+            for k, v in s.items():
+                if k == "atoms":
+                    # use to_hdf from Atoms class
+                    v.to_hdf(hdf = sub, group_name = "atoms")
+                else:
+                    # otherwise use the normal hdfio implementation
+                    sub[k] = v
+
+    def from_hdf(self, hdf = None, group_name = None):
+        super().from_hdf(hdf = hdf, group_name = group_name)
+
+        self.__structures.clear()
+
+        hdf = self.project_hdf5["structures"]
+        for i in sorted(hdf.list_groups()):
+            sub = hdf[i]
+
+            a = Atoms()
+            a.from_hdf(hdf = sub, group_name = "atoms")
+            meta = sub.list_groups()
+            meta.remove("atoms")
+
+            s = {"atoms": a}
+            self.__structures.append(s)
+            for k in meta:
+                s[k] = sub[k]
